requirements.txt
flask==2.3.3
python-dotenv==1.0.0
PyPDF2==3.0.1
google-generativeai==0.3.2
markupsafe==2.1.3
weasyprint==60.2
flask-login==0.6.3
flask-sqlalchemy==3.0.5
python-docx==0.8.11
gunicorn==21.2.0